requirements.txt
streamlit
spacy
scikit-learn
pdfminer.six
python-docx
joblib
numpy
pandas